separate internal/project package lib functions
